Refactor: Simplify semantic and codegen modules by merging small files
- Merged `src/semantic/assembled.rs` into `src/semantic/assembler.rs`.
- Merged `src/codegen/types.rs` into `src/codegen/utils.rs`.
- Merged `src/semantic/types.rs` into `src/semantic/model.rs`.
- Deleted dead code: `to_rust_ownership`.
- Updated `src/codegen/mod.rs` to re-export `to_rust_type` for integration tests.
- Updated `.jules/razor.md`.
